Reorganize whole codebase
- use sbt-typelevel to setup ci
- use different projects for tests and codebase
- add some scaladocs
- revamp build.sbt
