+++ .scalafmt.conf
@@ -1,4 +1,4 @@
 version = 3.6.1
 runner.dialect = scala3
 align.preset = more
-maxColumn = 120
+maxColumn = 80
