+++ core/src/main/scala/net/tazato/Main.scala
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2023 Connor James Smith
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package net.tazato
 
 import cats.effect.*
@@ -11,6 +27,8 @@ import net.tazato.surreal.ql.statements.*
 import org.typelevel.log4cats.*
 import org.typelevel.log4cats.slf4j.Slf4jFactory
 
+/** Testbed for interacting with surrealdb
+  */
 object Main extends IOApp.Simple {
   implicit val logging: LoggerFactory[IO]   = Slf4jFactory.create[IO]
   val logger: SelfAwareStructuredLogger[IO] = LoggerFactory[IO].getLogger
@@ -22,10 +40,13 @@ object Main extends IOApp.Simple {
             Method.POST,
             uri = uri"http://localhost:8000/sql",
             headers = Headers(
-              "NS"            -> "baby",
-              "DB"            -> "tazato",
-              "Accept"        -> "application/json",
-              "Authorization" -> ("Basic " ++ BasicCredentials("root", "root").token)
+              "NS"     -> "baby",
+              "DB"     -> "tazato",
+              "Accept" -> "application/json",
+              "Authorization" -> ("Basic " ++ BasicCredentials(
+                "root",
+                "root"
+              ).token)
             )
           ).withEntity(
             Define.NAMESPACE("baby").render
