+++ core/src/main/scala/net/tazato/surreal/ql/types/Base.scala
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2023 Connor James Smith
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.tazato.surreal.ql.types
+
+/** Enum that captures keywords having to do with namespace or database
+  */
+enum Base(val value: String):
+  case DATABASE  extends Base("DATABASE")
+  case DB        extends Base("DB")
+  case NAMESPACE extends Base("NAMESPACE")
+  case NS        extends Base("NS")
