+++ core/src/main/scala/net/tazato/surreal/ql/types/Render.scala
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2023 Connor James Smith
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.tazato.surreal.ql.types
+
+/** Renders (makes into a string) the type
+  * @tparam any
+  *   type
+  */
+trait Render[-A]:
+
+  /** @param x
+    *   the thing to render
+    * @return
+    *   the rendered thing
+    */
+  def render(x: A): String
