--- src/main/scala/net/tazato/surreal/ql/types/Render.scala
@@ -1,4 +1,0 @@
-package net.tazato.surreal.ql.types
-
-trait Render[-A]:
-  def render(x: A): String
