+++ tests/src/test/scala/net/tazato/surreal/ql/statements/SelectSuite.scala
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2023 Connor James Smith
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package net.tazato.surreal.ql.statements
 import weaver._
 
@@ -5,7 +21,10 @@ object SelectSuite extends FunSuite {
   import Select.{Target => Targ, *}
   test("Select Statement: render works, simple case") {
     val expected = """SELECT * FROM person;"""
-    val rendered = SELECT(selectFields = Seq(Field("*")), targets = Seq(Targ("person"))).render
+    val rendered = SELECT(
+      selectFields = Seq(Field("*")),
+      targets = Seq(Targ("person"))
+    ).render
     expect.same(expected, rendered)
   }
 
@@ -24,7 +43,8 @@ object SelectSuite extends FunSuite {
     val expected = """SELECT name AS user_name, address FROM person;"""
     val rendered = Select
       .SELECT(
-        selectFields = Seq(Field("name", Some(Alias("user_name"))), Field("address")),
+        selectFields =
+          Seq(Field("name", Some(Alias("user_name"))), Field("address")),
         targets = Seq(Targ("person"))
       )
       .render
@@ -34,7 +54,11 @@ object SelectSuite extends FunSuite {
   test("Select Statement: render works, split clause") {
     val expected = """SELECT * FROM user SPLIT AT emails;"""
     val rendered =
-      SELECT(selectFields = Seq(Field("*")), targets = Seq(Targ("user")), splitBy = Some(Field("emails"))).render
+      SELECT(
+        selectFields = Seq(Field("*")),
+        targets = Seq(Targ("user")),
+        splitBy = Some(Field("emails"))
+      ).render
     expect.same(expected, rendered)
   }
